-moved useDynamicHeight.tsx & useEvents.tsx into lib
